Create BoardProvider to manage column and card state
A normalized key/value store was chosen to provide O(1) lookups for columns and cards. A column holds a list of cardIds so that it can map over them and look up cards in the store to display in the column. For now, how positioning works is undecided. It could be based on array positionining of the cardIds, or we may add an additional position prop.